requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
ecdsa==0.19.1
idna==3.10
pyasn1==0.4.8
pycparser==2.22
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
